+++ app/src/main/java/com/iot/wateranalyst/ui/main/WaterData.kt
@@ -1,5 +1,6 @@
 package com.iot.wateranalyst.ui.main
 
-class WaterData(val pH: Float, val hardness: Float, val solids: Float,  val chloramines: Float, val sulfate: Float,
-                val conductivity: Float, val organicCarbon: Float, val trihalomethanes: Float, val turbidity: Float) {
+class WaterData(
+    val pH: Double, val hardness: Double, val solids: Double, val chloramines: Double, val sulfate: Double,
+    val conductivity: Double, val organicCarbon: Double, val trihalomethanes: Double, val turbidity: Double) {
 }
